scrap_IPC/requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pandas==1.3.0
mysql-connector-python==8.0.26
openpyxl==3.0.7
#pip install xlrd